Add DataSourceJdbcUrlMetaData (#15267)
