+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/datasource/pool/metadata/type/dbcp/TomcatDBCPDataSourcePoolMetaData.java
@@ -15,53 +15,55 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.datasource.pool.metadata.fixture;
+package org.apache.shardingsphere.infra.datasource.pool.metadata.type.dbcp;
 
+import lombok.Getter;
 import org.apache.shardingsphere.infra.datasource.pool.metadata.DataSourcePoolMetaData;
-import org.apache.shardingsphere.test.mock.MockedDataSource;
+import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
+import java.util.LinkedList;
 import java.util.Map;
 
-public final class MockedDataSourcePoolMetaData implements DataSourcePoolMetaData<MockedDataSource> {
+/**
+ * Tomcat JDBC data source pool meta data.
+ */
+@Getter
+public final class TomcatDBCPDataSourcePoolMetaData implements DataSourcePoolMetaData<BasicDataSource> {
+    
+    private final Collection<String> transientFieldNames = new LinkedList<>();
+    
+    public TomcatDBCPDataSourcePoolMetaData() {
+        buildTransientFieldNames();
+    }
+    
+    private void buildTransientFieldNames() {
+        transientFieldNames.add("closed");
+    }
     
     @Override
     public Map<String, Object> getDefaultProperties() {
-        Map<String, Object> result = new HashMap<>(1, 1);
-        result.put("maxPoolSize", 100);
-        return result;
+        return Collections.emptyMap();
     }
     
     @Override
     public Map<String, Object> getInvalidProperties() {
-        Map<String, Object> result = new HashMap<>(2, 1);
-        result.put("maxPoolSize", -1);
-        result.put("minPoolSize", -1);
-        return result;
+        return Collections.emptyMap();
     }
     
     @Override
     public Map<String, String> getPropertySynonyms() {
-        Map<String, String> result = new HashMap<>(2, 1);
-        result.put("maxPoolSize", "maxPoolSize");
-        result.put("minPoolSize", "minPoolSize");
-        return result;
+        return Collections.emptyMap();
     }
     
     @Override
-    public Collection<String> getTransientFieldNames() {
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public MockedDataSourceJdbcUrlMetaData getJdbcUrlMetaData() {
-        return new MockedDataSourceJdbcUrlMetaData();
+    public TomcatDBCPDataSourceJdbcUrlMetaData getJdbcUrlMetaData() {
+        return new TomcatDBCPDataSourceJdbcUrlMetaData();
     }
     
     @Override
     public String getType() {
-        return MockedDataSource.class.getName();
+        return BasicDataSource.class.getName();
     }
 }
